+++ rust/operator-binary/Cargo.toml
@@ -15,6 +15,7 @@ anyhow.workspace = true
 clap.workspace = true
 fnv.workspace = true
 futures.workspace = true
+indoc.workspace = true
 pin-project.workspace = true
 rand.workspace = true
 semver.workspace = true
